[BUGFIX] initialize meta on the first local set
If the local value is set before it is accessed, it incorrectly returns
the remote value when it is accessed. This is due to the prevRemote
value not getting initialized until the getter is called, causing the
local set value to get ignored until the next get call.

The fix here is to initialize the prevRemote value on the first property
set call.