Add support for custom comparator function in localCopy
Similiar to `dedupeTracked()` supports, we now can pass a custom function to `localCopy()` to compare the previous remote value with the incoming one. This is useful when you deal with non-primitive values, where a strict equal check will always return false while you still might want to consider the (complex) values as equal and thus not update the local copy.